+++ src/pages/piecemaker/timelines/index.js
@@ -1,5 +1,6 @@
 import create from './create'
 import edit from './edit'
+import exportTimeline from './export'
 import list from './list'
 import show from './sessions'
 import annotate from './annotate'
@@ -8,6 +9,7 @@ import user from './user'
 export {
   create,
   edit,
+  exportTimeline,
   list,
   show,
   annotate,