+++ src/router/routes/piecemaker.js
@@ -46,6 +46,12 @@ export default {
       meta: {private: true}
     },
     {
+      path: '/piecemaker/timelines/:uuid/export',
+      component: () => import('pages/piecemaker/timelines/export'),
+      name: 'piecemaker.timelines.export',
+      meta: {private: true}
+    },
+    {
       path: '/piecemaker/timelines/:uuid/search',
       component: () => import('pages/piecemaker/timelines/search'),
       name: 'piecemaker.timelines.search',